+++ packages/terra-time-input/tests/wdio/time-input-spec.js
@@ -576,6 +576,10 @@ describe('Time Input shortcut key operations', () => {
         const timevalue = currentDate.toISOString().split('T')[1].split('.')[0].split(':');
         expect($('#timeInput input[name="terra-time-hour-time-input"]')).toHaveValue(timevalue[0]);
         expect($('#timeInput input[name="terra-time-minute-time-input"]')).toHaveValue(timevalue[1]);
+        const minuteValue = Number($('#timeInput input[name="terra-time-minute-time-input"]').getValue());
+        const minuteTimeValue = Number(timevalue[1]);
+        const minuteInRange = (minuteValue === minuteTimeValue - 1 || minuteValue === minuteTimeValue);
+        expect(minuteInRange).toBe(true);
       });
 
       it('should set time to current time minus one minute if partially filled', () => {
@@ -651,7 +655,7 @@ describe('Time Input shortcut key operations', () => {
         expect($('#timeInput input[name="terra-time-minute-time-input"]')).toHaveValue(timevalue[1]);
         const secondsValue = Number($('#timeInput input[name="terra-time-minute-time-input"]').getValue());
         const secondsTimeValue = Number(timevalue[1]);
-        const secondsInRange = (secondsTimeValue === secondsValue + 1 || secondsTimeValue === secondsValue);
+        const secondsInRange = (secondsTimeValue === secondsValue - 1 || secondsTimeValue === secondsValue);
         expect(secondsInRange).toBe(true);
       });
 
@@ -671,7 +675,7 @@ describe('Time Input shortcut key operations', () => {
         expect($('#timeInput input[name="terra-time-minute-time-input"]')).toHaveValue(timevalue[1]);
         const secondsValue = Number($('#timeInput input[name="terra-time-minute-time-input"]').getValue());
         const secondsTimeValue = Number(timevalue[1]);
-        const secondsInRange = (secondsTimeValue === secondsValue + 1 || secondsTimeValue === secondsValue);
+        const secondsInRange = (secondsTimeValue === secondsValue - 1 || secondsTimeValue === secondsValue);
         expect(secondsInRange).toBe(true);
       });
 
@@ -691,7 +695,7 @@ describe('Time Input shortcut key operations', () => {
 
   describe('pressing +', () => {
     describe('in the hour input', () => {
-      it('should set time to current time plus one minute if blank', () => {
+      it('should set time to current time plus one hour if blank', () => {
         browser.url('/raw/tests/cerner-terra-framework-docs/time-input/time-input/default');
         browser.refresh();
         Terra.hideInputCaret('#timeInput input[name="terra-time-hour-time-input"]');
@@ -746,6 +750,10 @@ describe('Time Input shortcut key operations', () => {
         const timevalue = currentDate.toISOString().split('T')[1].split('.')[0].split(':');
         expect($('#timeInput input[name="terra-time-hour-time-input"]')).toHaveValue(timevalue[0]);
         expect($('#timeInput input[name="terra-time-minute-time-input"]')).toHaveValue(timevalue[1]);
+        const minuteValue = Number($('#timeInput input[name="terra-time-minute-time-input"]').getValue());
+        const minuteTimeValue = Number(timevalue[1]);
+        const minuteInRange = (minuteValue === minuteTimeValue + 1 || minuteValue === minuteTimeValue);
+        expect(minuteInRange).toBe(true);
       });
 
       it('should set time to current time plus one minute if partially filled', () => {