+++ packages/terra-framework-docs/CHANGELOG.md
@@ -8,6 +8,9 @@
 * Changed
   * Updated `terra-table` test example to support large text data.
 
+* Fixed
+  * Fixed label value when we have 12 hours time format in `terra-time-input` examples.
+
 ## 1.59.0 - (January 22, 2024)
 
 * Fixed
@@ -62,7 +65,7 @@
 * Changed
   * Updated examples and tests for expand all/collapse all for `terra-folder-tree`.
   * Updated `terra-theme-provider` examples to show the new ability to specify a theme density.
-  
+
 ## 1.52.0 - (December 13, 2023)
 
 * Changed