+++ packages/terra-framework-docs/CHANGELOG.md
@@ -1,9 +1,8 @@
 # Changelog
 
-* Changed
-  * Added `date-picker`, `time-input` and `date-time picker` tests to display new metadata.
-
 ## Unreleased
+
+## 1.60.0 - (January 25, 2024)
 
 * Added
   * Updated `date-input` examples to allow screen reader response for error message.
@@ -13,6 +12,9 @@
 
 * Fixed
   * Fixed label value when we have 12 hours time format in `terra-time-input` examples.
+
+* Changed
+  * Added `date-picker`, `time-input` and `date-time picker` tests to display new metadata.
 
 ## 1.59.0 - (January 22, 2024)